+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The changelog is available [on GitHub][2].
 * [#155](https://github.com/kowainik/relude/issues/155):
   Implement `Relude.Extra.Foldable` module.
 * Re-export `GHC.Float.atan2`.
+* Add `Monoid` and `Semigroup` instances for `Validation` type
 
 ## 0.5.0 — Mar 18, 2019